backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-extensions==3.1.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
